requirements.txt
Flask==2.2.2
pandas==1.3.3
fbprophet==0.7.1
plotly==5.3.1
matplotlib==3.4.3blinker==1.8.2
click==8.1.7
cmdstanpy==1.2.3
colorama==0.4.6
contourpy==1.2.1
convertdate==2.4.0
cycler==0.12.1
Cython==3.0.10
ephem==4.1.5
Flask==3.0.3
fonttools==4.53.0
holidays==0.50
importlib_metadata==7.1.0
importlib_resources==6.4.0
itsdangerous==2.2.0
Jinja2==3.1.4
kiwisolver==1.4.5
LunarCalendar==0.0.9
MarkupSafe==2.1.5
matplotlib==3.9.0
numpy==1.26.4
packaging==24.0
pandas==2.2.2
Pillow==8.3.1
plotly==5.22.0
prophet==1.1.5
pyarrow==16.1.0
PyMeeus==0.5.12
pyparsing==3.1.2
pystan==2.19.1.1
python-dateutil==2.9.0.post0
pytz==2024.1
setuptools-git==1.2
six==1.16.0
stanio==0.5.0
tenacity==8.3.0
tqdm==4.66.4
tzdata==2024.1
Werkzeug==3.0.3
zipp==3.19.2
